Clean up Aggregations in preparation for removing *args
